Created function helper base
