+++ src/engine/test/source/builder/builders/opBuilderHelperExists_test.cpp
@@ -7,9 +7,9 @@
  * Foundation.
  */
 
+#include <any>
 #include <gtest/gtest.h>
 #include <vector>
-#include <any>
 
 #include <baseTypes.hpp>
 
@@ -18,17 +18,19 @@
 using namespace base;
 namespace bld = builder::internals::builders;
 
-
 TEST(opBuilderHelperExists, Builds)
 {
-    auto tuple = std::make_tuple(std::string {"/field"}, std::vector<std::string> {});
+    auto tuple = std::make_tuple(
+        std::string {"/field"}, std::string {"exists"}, std::vector<std::string> {});
 
     ASSERT_NO_THROW(bld::opBuilderHelperExists(tuple));
 }
 
 TEST(opBuilderHelperExists, Exec_exists_false)
 {
-    auto tuple = std::make_tuple(std::string {"/field2check"}, std::vector<std::string> {});
+    auto tuple = std::make_tuple(std::string {"/field2check"},
+                                 std::string {"exists"},
+                                 std::vector<std::string> {});
 
     auto event1 = std::make_shared<json::Json>(R"({"fieldcheck": "valid"})");
 
@@ -41,7 +43,9 @@ TEST(opBuilderHelperExists, Exec_exists_false)
 
 TEST(opBuilderHelperExists, Exec_exists_true)
 {
-    auto tuple = std::make_tuple(std::string {"/field2check"}, std::vector<std::string> {});
+    auto tuple = std::make_tuple(std::string {"/field2check"},
+                                 std::string {"exists"},
+                                 std::vector<std::string> {});
 
     auto event1 = std::make_shared<json::Json>(R"({"field2check": "valid"})");
 
@@ -54,7 +58,9 @@ TEST(opBuilderHelperExists, Exec_exists_true)
 
 TEST(opBuilderHelperExists, Exec_multilevel_false)
 {
-    auto tuple = std::make_tuple(std::string {"/parentObjt_1/field2check"}, std::vector<std::string> {});
+    auto tuple = std::make_tuple(std::string {"/parentObjt_1/field2check"},
+                                 std::string {"exists"},
+                                 std::vector<std::string> {});
 
     auto event1 = std::make_shared<json::Json>(R"({
                     "parentObjt_2": {
@@ -76,7 +82,9 @@ TEST(opBuilderHelperExists, Exec_multilevel_false)
 
 TEST(opBuilderHelperExists, Exec_multilevel_true)
 {
-    auto tuple = std::make_tuple(std::string {"/parentObjt_1/field2check"}, std::vector<std::string> {});
+    auto tuple = std::make_tuple(std::string {"/parentObjt_1/field2check"},
+                                 std::string {"exists"},
+                                 std::vector<std::string> {});
 
     auto event1 = std::make_shared<json::Json>(R"({
                     "parentObjt_2": {
